# server/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"mongodb.ObjectId": {
				"type": "string",
				"description": "A class representation of the BSON ObjectId type."
			},
			"Product": {
				"properties": {
					"name": {
						"type": "string"
					},
					"shortDescription": {
						"type": "string"
					},
					"detailedDescription": {
						"type": "string"
					},
					"category": {
						"type": "string",
						"enum": [
							"Painting",
							"Sculptor",
							"Ornament"
						]
					},
					"startingPrice": {
						"type": "number",
						"format": "double"
					},
					"bidEndDate": {
						"type": "string",
						"format": "date-time"
					},
					"sellerId": {
						"type": "string"
					},
					"_id": {
						"$ref": "#/components/schemas/mongodb.ObjectId"
					}
				},
				"required": [
					"name",
					"shortDescription",
					"detailedDescription",
					"category",
					"startingPrice",
					"bidEndDate",
					"sellerId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Bid": {
				"properties": {
					"productId": {
						"type": "string"
					},
					"buyerName": {
						"type": "string"
					},
					"buyerId": {
						"type": "string"
					},
					"buyerEmail": {
						"type": "string"
					},
					"bidAmount": {
						"type": "number",
						"format": "double"
					},
					"requestedDate": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"productId",
					"buyerName",
					"buyerId",
					"buyerEmail",
					"bidAmount",
					"requestedDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductDetail": {
				"properties": {
					"product": {
						"$ref": "#/components/schemas/Product"
					},
					"bids": {
						"items": {
							"$ref": "#/components/schemas/Bid"
						},
						"type": "array"
					}
				},
				"required": [
					"product",
					"bids"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "server",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/ping": {
			"get": {
				"operationId": "GetPing",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/product": {
			"get": {
				"operationId": "GetProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Product"
									},
									"type": "array",
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Retrieves the list  of a products.",
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "SaveProduct",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"description": "Save a product.",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Product"
							}
						}
					}
				}
			}
		},
		"/product/{productId}": {
			"get": {
				"operationId": "GetProductDetail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductDetail"
								}
							}
						}
					}
				},
				"description": "Retrieves the list  a products.",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}